alterações na infraestrutura

+++ infra/backend.tf
@@ -1,24 +1,41 @@
 resource "aws_instance" "api_server" {
     ami                     = var.ami_id
     instance_type           = var.instance_type
-    subnet_id               = aws_subnet.app_subnet.id 
-    vpc_security_group_ids  = [aws_security_group.backendSG.id] 
+    security_groups         = [aws_security_group.ServerEC_Backend.name] 
     key_name                = var.key_name
-    associate_public_ip_address = true
 
     tags = {
-        Name = "Backend-Instance"
+        Name = "E-Commerce - Backend"
     }
 
     user_data = <<-EOF
-                 #!/bin/bash
+                #!/bin/bash
+                exec > /var/log/user-data.log 2>&1  # Redirect output to log file
+                set -x  # Enable debugging
+
+                echo "Starting backend server setup..."
+
                 cd /home/ubuntu
-                git clone https://github.com/InesMRibeiro/E_Commerce_App.git
+                if [ ! -d "E_Commerce_App" ]; then
+                    git clone https://github.com/InesMRibeiro/E_Commerce_App.git
+                else
+                    echo "Repository already cloned."
+                fi
+
                 sudo apt update -y
                 sudo apt install -y python3 python3-pip python3-venv git
+
                 python3 -m venv venv
                 source venv/bin/activate
+
                 cd E_Commerce_App/backend
-                pip install -r requirements.txt
+
+                if [ -f "requirements.txt" ]; then
+                     pip install -r requirements.txt
+                else
+                     echo "requirements.txt not found!" >> /var/log/user-data.log
+                fi
+                echo "Backend server setup completed!"
+                
                 EOF
 }
